@@ -1,56 +1,36 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp1.1</TargetFramework>
-    <PackageTargetFallback>$(PackageTargetFallback);portable-net45+win8+wp8+wpa81;</PackageTargetFallback>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <AssetTargetFallback>$(AssetTargetFallback);portable-net45+win8+wp8+wpa81;</AssetTargetFallback>
+
     <GenerateDocumentationFile>false</GenerateDocumentationFile>
     <Configurations>Debug;Release;VS</Configurations>
     <TypeScriptToolsVersion>2.8</TypeScriptToolsVersion>
-
+    <SignAssembly>false</SignAssembly>
+    <UserSecretsId>f7c4472d-ea58-47e2-9129-489edcd48abe</UserSecretsId>
   </PropertyGroup>
-
-  <ItemGroup>
-    <Content Remove="wwwroot\lib\monaco-editor\monaco.d.ts" />
-  </ItemGroup>
 
   <ItemGroup>
     <None Remove="Properties\PublishProfiles\protogensite20170521042027 - Web Deploy.pubxml" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore" Version="1.1.2" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="1.1.3" />
-    <PackageReference Include="Microsoft.AspNetCore.StaticFiles" Version="1.1.2" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="1.1.2" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.BrowserLink" Version="1.1.2" />
+    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="3.2.1" />
+    <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.16.0" />
+    <PackageReference Include="System.Collections.Immutable" Version="1.7.1" />
+    <PackageReference Include="System.Text.Json" Version="5.0.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
 
   <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="1.0.1" />
+    <ProjectReference Include="..\protobuf-net\protobuf-net.csproj" />
+    <ProjectReference Include="..\protobuf-net.Reflection\protobuf-net.Reflection.csproj" />
+    <ProjectReference Include="..\protogen.site.blazor.client\protogen.site.blazor.client.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <None Include="wwwroot\protoc\google\protobuf\duration.proto" />
-    <None Include="wwwroot\protoc\google\protobuf\timestamp.proto" />
-  </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\protobuf-net\aqlaserializer.csproj" />
-    <ProjectReference Include="..\protobuf-net.Reflection\aqlaserializer.Reflection.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <TypeScriptCompile Include="wwwroot\lib\monaco-editor\monaco.d.ts" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update=".npmrc">
-      <DependentUpon>package.json</DependentUpon>
-    </None>
-    <None Update="postinstall.js">
-      <DependentUpon>package.json</DependentUpon>
-    </None>
     <None Update="protoc.exe">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
