@@ -1,15 +1,24 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <VersionPrefix>$(ProtoGenVersion)</VersionPrefix>
-    <!--<VersionSuffix></VersionSuffix>-->
     <OutputType>Exe</OutputType>
-    <TargetFrameworks>net462;netcoreapp2.0</TargetFrameworks>
+    <TargetFrameworks>net462;netcoreapp3.1;net5.0</TargetFrameworks>
     <Configurations>Debug;Release;VS</Configurations>
     <AssemblyName>protogen</AssemblyName>
     <Title>protobuf-net command-line "global tool" for .NET code-generation from .proto schema files</Title>
+    <GrpcTools>false</GrpcTools>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(TargetFramework)'=='netcoreapp3.1' or '$(TargetFramework)'=='net5.0'">
+    <GrpcTools>true</GrpcTools>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(GrpcTools)'=='true'">
+    <DefineConstants>$(DefineConstants);GRPC_TOOLS</DefineConstants>
   </PropertyGroup>
   <ItemGroup>
-    <ProjectReference Include="../protobuf-net/aqlaserializer.csproj" />
-    <ProjectReference Include="../protobuf-net.Reflection/aqlaserializer.Reflection.csproj" />
+    <ProjectReference Include="../protobuf-net/protobuf-net.csproj" />
+    <ProjectReference Include="../protobuf-net.Reflection/protobuf-net.Reflection.csproj" />
+    <PackageReference Include="System.Collections.Immutable" Version="1.7.1" />
+    <PackageReference Include="protobuf-net.Grpc.Reflection" Version="1.0.123" Condition="'$(GrpcTools)'=='true'" />
+    <PackageReference Include="Grpc.Net.Client" Version="2.33.1" Condition="'$(GrpcTools)'=='true'"/>
   </ItemGroup>
 </Project>
