@@ -2,7 +2,8 @@
 // Consider using 'partial classes' to extend these types
 // Input: my.proto
 
-#pragma warning disable CS1591, CS0612, CS3021, IDE1006
+#pragma warning disable IDE0079
+#pragma warning disable CS1591, CS0612, CS0618, CS3021, IDE1006
 namespace protogen
 {
 
@@ -105,4 +106,4 @@ namespace protogen
 
 }
 
-#pragma warning restore CS1591, CS0612, CS3021, IDE1006
+#pragma warning restore IDE0079, CS1591, CS0612, CS0618, CS3021, IDE1006
