@@ -1,27 +1,29 @@
 ﻿using BenchmarkDotNet.Attributes;
-using AqlaSerializer;
-using AqlaSerializer.Meta;
+using BenchmarkDotNet.Jobs;
+using ProtoBuf;
+using ProtoBuf.Meta;
 using System;
 using System.Buffers;
 using System.IO;
 
+#if NEW_API
 namespace Benchmark
 {
-    [ClrJob, CoreJob, MemoryDiagnoser]
+    [SimpleJob(RuntimeMoniker.Net472), SimpleJob(RuntimeMoniker.NetCoreApp31), SimpleJob(RuntimeMoniker.NetCoreApp50), MemoryDiagnoser]
     public class SpanPerformance
     {
         private MemoryStream _ms;
         private ReadOnlySequence<byte> _ros;
-        public ProtoReader ReadMS(out ProtoReader.State state)
-            => ProtoReader.Create(out state, _ms, Model);
+        public ProtoReader.State ReadMS()
+            => ProtoReader.State.Create(_ms, Model);
 
-        public ProtoReader ReadROS(out ProtoReader.State state)
-            => ProtoReader.Create(out state, _ros, Model);
+        public ProtoReader.State ReadROS()
+            => ProtoReader.State.Create(_ros, Model);
 
-        public ProtoReader ReadROM(out ProtoReader.State state)
+        public ProtoReader.State ReadROM()
         {
             if (!_ros.IsSingleSegment) throw new InvalidOperationException("Expected single segment");
-            return ProtoReader.Create(out state, _ros.First, Model);
+            return ProtoReader.State.Create(_ros.First, Model);
         }
 
         public TypeModel Model => RuntimeTypeModel.Default;
@@ -38,30 +40,25 @@ namespace Benchmark
         public void MemoryStream()
         {
             _ms.Position = 0;
-            using (var reader = ReadMS(out var state))
-            {
-                var dal = Model.Deserialize(reader, ref state, null, typeof(DAL.Database));
-                GC.KeepAlive(dal);
-            }
+            using var reader = ReadMS();
+            var dal = reader.DeserializeRoot<protogen.Database>();
+            GC.KeepAlive(dal);
         }
 
         [Benchmark]
         public void ReadOnlySequence()
         {
-            using (var reader = ReadROS(out var state))
-            {
-                var dal = Model.Deserialize(reader, ref state, null, typeof(DAL.Database));
-                GC.KeepAlive(dal);
-            }
+            using var reader = ReadROS();
+            var dal = reader.DeserializeRoot<protogen.Database>();
+            GC.KeepAlive(dal);
         }
         [Benchmark]
         public void ReadOnlyMemory()
         {
-            using (var reader = ReadROM(out var state))
-            {
-                var dal = Model.Deserialize(reader, ref state, null, typeof(DAL.Database));
-                GC.KeepAlive(dal);
-            }
+            using var reader = ReadROM();
+            var dal = reader.DeserializeRoot<protogen.Database>();
+            GC.KeepAlive(dal);
         }
     }
 }
+#endif
