@@ -19,9 +19,10 @@ namespace AqlaSerializer.Build
 
     class Program
     {
-
-        static void Main(string[] args)
-        {
-        }
+        #pragma warning disable IDE0060
+                static void Main(string[] args)
+        #pragma warning restore IDE0060
+                {
+                }
     }
 }
