@@ -73,6 +73,7 @@ namespace AqlaSerializer.Reflection
                         break;
                 }
             }
+            [System.Diagnostics.CodeAnalysis.SuppressMessage("Style", "IDE0066:Convert switch statement to expression", Justification = "Readability")]
             internal string GetStorage(FieldDescriptorProto.Type type, string typeName)
             {
                 switch (type)
@@ -99,17 +100,13 @@ namespace AqlaSerializer.Reflection
                     case FieldDescriptorProto.Type.TypeUint64:
                         return "UInt64";
                     case FieldDescriptorProto.Type.TypeMessage:
-                        switch (typeName)
+                        return typeName switch
                         {
-                            case ".google.protobuf.Timestamp":
-                                return "DateTime";
-                            case ".google.protobuf.Duration":
-                                return "TimeSpan";
-                            case ".bcl.Guid":
-                                return "Guid";
-                            default:
-                                return "Object";
-                        }
+                            ".google.protobuf.Timestamp" => "DateTime",
+                            ".google.protobuf.Duration" => "TimeSpan",
+                            ".bcl.Guid" => "Guid",
+                            _ => "Object",
+                        };
                     default:
                         return "Object";
                 }
