@@ -1,8 +1,14 @@
-// This file was generated by a tool; you should avoid making direct changes.
-// Consider using 'partial classes' to extend these types
-// Input: protogen.proto
+// <auto-generated>
+//   This file was generated by a tool; you should avoid making direct changes.
+//   Consider using 'partial classes' to extend these types
+//   Input: my.proto
+// </auto-generated>
 
-#pragma warning disable CS0612, CS1591, CS3021, IDE1006, RCS1036, RCS1057, RCS1085, RCS1192
+#region Designer generated code
+#pragma warning disable CS0612, CS0618, CS1591, CS3021, IDE1006, RCS1036, RCS1057, RCS1085, RCS1192
+using ProtoBuf.Meta;
+using ProtoBuf.Reflection.Internal;
+
 namespace AqlaSerializer.Reflection
 {
 
@@ -58,6 +64,10 @@ namespace AqlaSerializer.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string ExtensionTypeName { get; set; } = "";
 
+        [global::ProtoBuf.ProtoMember(4, Name = @"namespace")]
+        [global::System.ComponentModel.DefaultValue("")]
+        public string Namespace { get; set; } = "";
+
     }
 
     [global::ProtoBuf.ProtoContract()]
@@ -95,6 +105,10 @@ namespace AqlaSerializer.Reflection
 
         [global::ProtoBuf.ProtoMember(2, Name = @"access")]
         public Access Access { get; set; }
+
+        [global::ProtoBuf.ProtoMember(3, Name = @"namespace")]
+        [global::System.ComponentModel.DefaultValue("")]
+        public string Namespace { get; set; } = "";
 
     }
 
@@ -169,57 +183,59 @@ namespace AqlaSerializer.Reflection
         Internal = 3,
     }
 
-    public static class Extensions
+    public static partial class Extensions
     {
+        private static TypeModel Serializer => CustomProtogenSerializer.Instance;
         public static ProtogenFileOptions GetOptions(this global::Google.Protobuf.Reflection.FileOptions obj)
-            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenFileOptions>(obj, 1037);
+            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenFileOptions>(Serializer, obj, 1037);
 
         public static void SetOptions(this global::Google.Protobuf.Reflection.FileOptions obj, ProtogenFileOptions value)
-            => global::ProtoBuf.Extensible.AppendValue<ProtogenFileOptions>(obj, 1037, value);
+            => global::ProtoBuf.Extensible.AppendValue<ProtogenFileOptions>(Serializer, obj, 1037, value);
 
         public static ProtogenMessageOptions GetOptions(this global::Google.Protobuf.Reflection.MessageOptions obj)
-            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenMessageOptions>(obj, 1037);
+            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenMessageOptions>(Serializer, obj, 1037);
 
         public static void SetOptions(this global::Google.Protobuf.Reflection.MessageOptions obj, ProtogenMessageOptions value)
-            => global::ProtoBuf.Extensible.AppendValue<ProtogenMessageOptions>(obj, 1037, value);
+            => global::ProtoBuf.Extensible.AppendValue<ProtogenMessageOptions>(Serializer, obj, 1037, value);
 
         public static ProtogenFieldOptions GetOptions(this global::Google.Protobuf.Reflection.FieldOptions obj)
-            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenFieldOptions>(obj, 1037);
+            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenFieldOptions>(Serializer, obj, 1037);
 
         public static void SetOptions(this global::Google.Protobuf.Reflection.FieldOptions obj, ProtogenFieldOptions value)
-            => global::ProtoBuf.Extensible.AppendValue<ProtogenFieldOptions>(obj, 1037, value);
+            => global::ProtoBuf.Extensible.AppendValue<ProtogenFieldOptions>(Serializer, obj, 1037, value);
 
         public static ProtogenEnumOptions GetOptions(this global::Google.Protobuf.Reflection.EnumOptions obj)
-            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenEnumOptions>(obj, 1037);
+            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenEnumOptions>(Serializer, obj, 1037);
 
         public static void SetOptions(this global::Google.Protobuf.Reflection.EnumOptions obj, ProtogenEnumOptions value)
-            => global::ProtoBuf.Extensible.AppendValue<ProtogenEnumOptions>(obj, 1037, value);
+            => global::ProtoBuf.Extensible.AppendValue<ProtogenEnumOptions>(Serializer, obj, 1037, value);
 
         public static ProtogenEnumValueOptions GetOptions(this global::Google.Protobuf.Reflection.EnumValueOptions obj)
-            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenEnumValueOptions>(obj, 1037);
+            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenEnumValueOptions>(Serializer, obj, 1037);
 
         public static void SetOptions(this global::Google.Protobuf.Reflection.EnumValueOptions obj, ProtogenEnumValueOptions value)
-            => global::ProtoBuf.Extensible.AppendValue<ProtogenEnumValueOptions>(obj, 1037, value);
+            => global::ProtoBuf.Extensible.AppendValue<ProtogenEnumValueOptions>(Serializer, obj, 1037, value);
 
         public static ProtogenServiceOptions GetOptions(this global::Google.Protobuf.Reflection.ServiceOptions obj)
-            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenServiceOptions>(obj, 1037);
+            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenServiceOptions>(Serializer, obj, 1037);
 
         public static void SetOptions(this global::Google.Protobuf.Reflection.ServiceOptions obj, ProtogenServiceOptions value)
-            => global::ProtoBuf.Extensible.AppendValue<ProtogenServiceOptions>(obj, 1037, value);
+            => global::ProtoBuf.Extensible.AppendValue<ProtogenServiceOptions>(Serializer, obj, 1037, value);
 
         public static ProtogenMethodOptions GetOptions(this global::Google.Protobuf.Reflection.MethodOptions obj)
-            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenMethodOptions>(obj, 1037);
+            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenMethodOptions>(Serializer, obj, 1037);
 
         public static void SetOptions(this global::Google.Protobuf.Reflection.MethodOptions obj, ProtogenMethodOptions value)
-            => global::ProtoBuf.Extensible.AppendValue<ProtogenMethodOptions>(obj, 1037, value);
+            => global::ProtoBuf.Extensible.AppendValue<ProtogenMethodOptions>(Serializer, obj, 1037, value);
 
         public static ProtogenOneofOptions GetOptions(this global::Google.Protobuf.Reflection.OneofOptions obj)
-            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenOneofOptions>(obj, 1037);
+            => obj == null ? default : global::ProtoBuf.Extensible.GetValue<ProtogenOneofOptions>(Serializer, obj, 1037);
 
         public static void SetOptions(this global::Google.Protobuf.Reflection.OneofOptions obj, ProtogenOneofOptions value)
-            => global::ProtoBuf.Extensible.AppendValue<ProtogenOneofOptions>(obj, 1037, value);
+            => global::ProtoBuf.Extensible.AppendValue<ProtogenOneofOptions>(Serializer, obj, 1037, value);
 
     }
 }
 
-#pragma warning restore CS0612, CS1591, CS3021, IDE1006, RCS1036, RCS1057, RCS1085, RCS1192
+#pragma warning restore CS0612, CS0618, CS1591, CS3021, IDE1006, RCS1036, RCS1057, RCS1085, RCS1192
+#endregion
