@@ -1,8 +1,11 @@
-// This file was generated by a tool; you should avoid making direct changes.
-// Consider using 'partial classes' to extend these types
-// Input: descriptor.proto
+// <auto-generated>
+//   This file was generated by a tool; you should avoid making direct changes.
+//   Consider using 'partial classes' to extend these types
+//   Input: descriptor.proto
+// </auto-generated>
 
-#pragma warning disable CS0612, CS1591, CS3021, IDE1006, RCS1036, RCS1057, RCS1085, RCS1192
+#region Designer generated code
+#pragma warning disable CS0612, CS0618, CS1591, CS3021, IDE1006, RCS1036, RCS1057, RCS1085, RCS1192
 namespace Google.Protobuf.Reflection
 {
 
@@ -29,8 +32,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Name
         {
-            get { return __pbn__Name ?? ""; }
-            set { __pbn__Name = value; }
+            get => __pbn__Name ?? "";
+            set => __pbn__Name = value;
         }
         public bool ShouldSerializeName() => __pbn__Name != null;
         public void ResetName() => __pbn__Name = null;
@@ -40,8 +43,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Package
         {
-            get { return __pbn__Package ?? ""; }
-            set { __pbn__Package = value; }
+            get => __pbn__Package ?? "";
+            set => __pbn__Package = value;
         }
         public bool ShouldSerializePackage() => __pbn__Package != null;
         public void ResetPackage() => __pbn__Package = null;
@@ -78,8 +81,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Syntax
         {
-            get { return __pbn__Syntax ?? ""; }
-            set { __pbn__Syntax = value; }
+            get => __pbn__Syntax ?? "";
+            set => __pbn__Syntax = value;
         }
         public bool ShouldSerializeSyntax() => __pbn__Syntax != null;
         public void ResetSyntax() => __pbn__Syntax = null;
@@ -98,8 +101,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Name
         {
-            get { return __pbn__Name ?? ""; }
-            set { __pbn__Name = value; }
+            get => __pbn__Name ?? "";
+            set => __pbn__Name = value;
         }
         public bool ShouldSerializeName() => __pbn__Name != null;
         public void ResetName() => __pbn__Name = null;
@@ -142,8 +145,8 @@ namespace Google.Protobuf.Reflection
             [global::ProtoBuf.ProtoMember(1, Name = @"start")]
             public int Start
             {
-                get { return __pbn__Start.GetValueOrDefault(); }
-                set { __pbn__Start = value; }
+                get => __pbn__Start.GetValueOrDefault();
+                set => __pbn__Start = value;
             }
             public bool ShouldSerializeStart() => __pbn__Start != null;
             public void ResetStart() => __pbn__Start = null;
@@ -152,8 +155,8 @@ namespace Google.Protobuf.Reflection
             [global::ProtoBuf.ProtoMember(2, Name = @"end")]
             public int End
             {
-                get { return __pbn__End.GetValueOrDefault(); }
-                set { __pbn__End = value; }
+                get => __pbn__End.GetValueOrDefault();
+                set => __pbn__End = value;
             }
             public bool ShouldSerializeEnd() => __pbn__End != null;
             public void ResetEnd() => __pbn__End = null;
@@ -174,8 +177,8 @@ namespace Google.Protobuf.Reflection
             [global::ProtoBuf.ProtoMember(1, Name = @"start")]
             public int Start
             {
-                get { return __pbn__Start.GetValueOrDefault(); }
-                set { __pbn__Start = value; }
+                get => __pbn__Start.GetValueOrDefault();
+                set => __pbn__Start = value;
             }
             public bool ShouldSerializeStart() => __pbn__Start != null;
             public void ResetStart() => __pbn__Start = null;
@@ -184,8 +187,8 @@ namespace Google.Protobuf.Reflection
             [global::ProtoBuf.ProtoMember(2, Name = @"end")]
             public int End
             {
-                get { return __pbn__End.GetValueOrDefault(); }
-                set { __pbn__End = value; }
+                get => __pbn__End.GetValueOrDefault();
+                set => __pbn__End = value;
             }
             public bool ShouldSerializeEnd() => __pbn__End != null;
             public void ResetEnd() => __pbn__End = null;
@@ -218,8 +221,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Name
         {
-            get { return __pbn__Name ?? ""; }
-            set { __pbn__Name = value; }
+            get => __pbn__Name ?? "";
+            set => __pbn__Name = value;
         }
         public bool ShouldSerializeName() => __pbn__Name != null;
         public void ResetName() => __pbn__Name = null;
@@ -228,8 +231,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(3, Name = @"number")]
         public int Number
         {
-            get { return __pbn__Number.GetValueOrDefault(); }
-            set { __pbn__Number = value; }
+            get => __pbn__Number.GetValueOrDefault();
+            set => __pbn__Number = value;
         }
         public bool ShouldSerializeNumber() => __pbn__Number != null;
         public void ResetNumber() => __pbn__Number = null;
@@ -239,8 +242,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(Label.LabelOptional)]
         public Label label
         {
-            get { return __pbn__label ?? Label.LabelOptional; }
-            set { __pbn__label = value; }
+            get => __pbn__label ?? Label.LabelOptional;
+            set => __pbn__label = value;
         }
         public bool ShouldSerializelabel() => __pbn__label != null;
         public void Resetlabel() => __pbn__label = null;
@@ -250,8 +253,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(Type.TypeDouble)]
         public Type type
         {
-            get { return __pbn__type ?? Type.TypeDouble; }
-            set { __pbn__type = value; }
+            get => __pbn__type ?? Type.TypeDouble;
+            set => __pbn__type = value;
         }
         public bool ShouldSerializetype() => __pbn__type != null;
         public void Resettype() => __pbn__type = null;
@@ -261,8 +264,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string TypeName
         {
-            get { return __pbn__TypeName ?? ""; }
-            set { __pbn__TypeName = value; }
+            get => __pbn__TypeName ?? "";
+            set => __pbn__TypeName = value;
         }
         public bool ShouldSerializeTypeName() => __pbn__TypeName != null;
         public void ResetTypeName() => __pbn__TypeName = null;
@@ -272,8 +275,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Extendee
         {
-            get { return __pbn__Extendee ?? ""; }
-            set { __pbn__Extendee = value; }
+            get => __pbn__Extendee ?? "";
+            set => __pbn__Extendee = value;
         }
         public bool ShouldSerializeExtendee() => __pbn__Extendee != null;
         public void ResetExtendee() => __pbn__Extendee = null;
@@ -283,8 +286,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string DefaultValue
         {
-            get { return __pbn__DefaultValue ?? ""; }
-            set { __pbn__DefaultValue = value; }
+            get => __pbn__DefaultValue ?? "";
+            set => __pbn__DefaultValue = value;
         }
         public bool ShouldSerializeDefaultValue() => __pbn__DefaultValue != null;
         public void ResetDefaultValue() => __pbn__DefaultValue = null;
@@ -293,8 +296,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(9, Name = @"oneof_index")]
         public int OneofIndex
         {
-            get { return __pbn__OneofIndex.GetValueOrDefault(); }
-            set { __pbn__OneofIndex = value; }
+            get => __pbn__OneofIndex.GetValueOrDefault();
+            set => __pbn__OneofIndex = value;
         }
         public bool ShouldSerializeOneofIndex() => __pbn__OneofIndex != null;
         public void ResetOneofIndex() => __pbn__OneofIndex = null;
@@ -304,8 +307,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string JsonName
         {
-            get { return __pbn__JsonName ?? ""; }
-            set { __pbn__JsonName = value; }
+            get => __pbn__JsonName ?? "";
+            set => __pbn__JsonName = value;
         }
         public bool ShouldSerializeJsonName() => __pbn__JsonName != null;
         public void ResetJsonName() => __pbn__JsonName = null;
@@ -313,6 +316,16 @@ namespace Google.Protobuf.Reflection
 
         [global::ProtoBuf.ProtoMember(8, Name = @"options")]
         public FieldOptions Options { get; set; }
+
+        [global::ProtoBuf.ProtoMember(17, Name = @"proto3_optional")]
+        public bool Proto3Optional
+        {
+            get => __pbn__Proto3Optional.GetValueOrDefault();
+            set => __pbn__Proto3Optional = value;
+        }
+        public bool ShouldSerializeProto3Optional() => __pbn__Proto3Optional != null;
+        public void ResetProto3Optional() => __pbn__Proto3Optional = null;
+        private bool? __pbn__Proto3Optional;
 
         [global::ProtoBuf.ProtoContract()]
         public enum Type
@@ -379,8 +392,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Name
         {
-            get { return __pbn__Name ?? ""; }
-            set { __pbn__Name = value; }
+            get => __pbn__Name ?? "";
+            set => __pbn__Name = value;
         }
         public bool ShouldSerializeName() => __pbn__Name != null;
         public void ResetName() => __pbn__Name = null;
@@ -402,8 +415,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Name
         {
-            get { return __pbn__Name ?? ""; }
-            set { __pbn__Name = value; }
+            get => __pbn__Name ?? "";
+            set => __pbn__Name = value;
         }
         public bool ShouldSerializeName() => __pbn__Name != null;
         public void ResetName() => __pbn__Name = null;
@@ -431,8 +444,8 @@ namespace Google.Protobuf.Reflection
             [global::ProtoBuf.ProtoMember(1, Name = @"start")]
             public int Start
             {
-                get { return __pbn__Start.GetValueOrDefault(); }
-                set { __pbn__Start = value; }
+                get => __pbn__Start.GetValueOrDefault();
+                set => __pbn__Start = value;
             }
             public bool ShouldSerializeStart() => __pbn__Start != null;
             public void ResetStart() => __pbn__Start = null;
@@ -441,8 +454,8 @@ namespace Google.Protobuf.Reflection
             [global::ProtoBuf.ProtoMember(2, Name = @"end")]
             public int End
             {
-                get { return __pbn__End.GetValueOrDefault(); }
-                set { __pbn__End = value; }
+                get => __pbn__End.GetValueOrDefault();
+                set => __pbn__End = value;
             }
             public bool ShouldSerializeEnd() => __pbn__End != null;
             public void ResetEnd() => __pbn__End = null;
@@ -463,8 +476,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Name
         {
-            get { return __pbn__Name ?? ""; }
-            set { __pbn__Name = value; }
+            get => __pbn__Name ?? "";
+            set => __pbn__Name = value;
         }
         public bool ShouldSerializeName() => __pbn__Name != null;
         public void ResetName() => __pbn__Name = null;
@@ -473,8 +486,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(2, Name = @"number")]
         public int Number
         {
-            get { return __pbn__Number.GetValueOrDefault(); }
-            set { __pbn__Number = value; }
+            get => __pbn__Number.GetValueOrDefault();
+            set => __pbn__Number = value;
         }
         public bool ShouldSerializeNumber() => __pbn__Number != null;
         public void ResetNumber() => __pbn__Number = null;
@@ -496,8 +509,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Name
         {
-            get { return __pbn__Name ?? ""; }
-            set { __pbn__Name = value; }
+            get => __pbn__Name ?? "";
+            set => __pbn__Name = value;
         }
         public bool ShouldSerializeName() => __pbn__Name != null;
         public void ResetName() => __pbn__Name = null;
@@ -522,8 +535,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string Name
         {
-            get { return __pbn__Name ?? ""; }
-            set { __pbn__Name = value; }
+            get => __pbn__Name ?? "";
+            set => __pbn__Name = value;
         }
         public bool ShouldSerializeName() => __pbn__Name != null;
         public void ResetName() => __pbn__Name = null;
@@ -533,8 +546,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string InputType
         {
-            get { return __pbn__InputType ?? ""; }
-            set { __pbn__InputType = value; }
+            get => __pbn__InputType ?? "";
+            set => __pbn__InputType = value;
         }
         public bool ShouldSerializeInputType() => __pbn__InputType != null;
         public void ResetInputType() => __pbn__InputType = null;
@@ -544,8 +557,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string OutputType
         {
-            get { return __pbn__OutputType ?? ""; }
-            set { __pbn__OutputType = value; }
+            get => __pbn__OutputType ?? "";
+            set => __pbn__OutputType = value;
         }
         public bool ShouldSerializeOutputType() => __pbn__OutputType != null;
         public void ResetOutputType() => __pbn__OutputType = null;
@@ -558,8 +571,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool ClientStreaming
         {
-            get { return __pbn__ClientStreaming ?? false; }
-            set { __pbn__ClientStreaming = value; }
+            get => __pbn__ClientStreaming ?? false;
+            set => __pbn__ClientStreaming = value;
         }
         public bool ShouldSerializeClientStreaming() => __pbn__ClientStreaming != null;
         public void ResetClientStreaming() => __pbn__ClientStreaming = null;
@@ -569,8 +582,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool ServerStreaming
         {
-            get { return __pbn__ServerStreaming ?? false; }
-            set { __pbn__ServerStreaming = value; }
+            get => __pbn__ServerStreaming ?? false;
+            set => __pbn__ServerStreaming = value;
         }
         public bool ShouldSerializeServerStreaming() => __pbn__ServerStreaming != null;
         public void ResetServerStreaming() => __pbn__ServerStreaming = null;
@@ -589,8 +602,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string JavaPackage
         {
-            get { return __pbn__JavaPackage ?? ""; }
-            set { __pbn__JavaPackage = value; }
+            get => __pbn__JavaPackage ?? "";
+            set => __pbn__JavaPackage = value;
         }
         public bool ShouldSerializeJavaPackage() => __pbn__JavaPackage != null;
         public void ResetJavaPackage() => __pbn__JavaPackage = null;
@@ -600,8 +613,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string JavaOuterClassname
         {
-            get { return __pbn__JavaOuterClassname ?? ""; }
-            set { __pbn__JavaOuterClassname = value; }
+            get => __pbn__JavaOuterClassname ?? "";
+            set => __pbn__JavaOuterClassname = value;
         }
         public bool ShouldSerializeJavaOuterClassname() => __pbn__JavaOuterClassname != null;
         public void ResetJavaOuterClassname() => __pbn__JavaOuterClassname = null;
@@ -611,8 +624,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool JavaMultipleFiles
         {
-            get { return __pbn__JavaMultipleFiles ?? false; }
-            set { __pbn__JavaMultipleFiles = value; }
+            get => __pbn__JavaMultipleFiles ?? false;
+            set => __pbn__JavaMultipleFiles = value;
         }
         public bool ShouldSerializeJavaMultipleFiles() => __pbn__JavaMultipleFiles != null;
         public void ResetJavaMultipleFiles() => __pbn__JavaMultipleFiles = null;
@@ -622,8 +635,8 @@ namespace Google.Protobuf.Reflection
         [global::System.Obsolete]
         public bool JavaGenerateEqualsAndHash
         {
-            get { return __pbn__JavaGenerateEqualsAndHash.GetValueOrDefault(); }
-            set { __pbn__JavaGenerateEqualsAndHash = value; }
+            get => __pbn__JavaGenerateEqualsAndHash.GetValueOrDefault();
+            set => __pbn__JavaGenerateEqualsAndHash = value;
         }
         public bool ShouldSerializeJavaGenerateEqualsAndHash() => __pbn__JavaGenerateEqualsAndHash != null;
         public void ResetJavaGenerateEqualsAndHash() => __pbn__JavaGenerateEqualsAndHash = null;
@@ -633,8 +646,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool JavaStringCheckUtf8
         {
-            get { return __pbn__JavaStringCheckUtf8 ?? false; }
-            set { __pbn__JavaStringCheckUtf8 = value; }
+            get => __pbn__JavaStringCheckUtf8 ?? false;
+            set => __pbn__JavaStringCheckUtf8 = value;
         }
         public bool ShouldSerializeJavaStringCheckUtf8() => __pbn__JavaStringCheckUtf8 != null;
         public void ResetJavaStringCheckUtf8() => __pbn__JavaStringCheckUtf8 = null;
@@ -644,8 +657,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(OptimizeMode.Speed)]
         public OptimizeMode OptimizeFor
         {
-            get { return __pbn__OptimizeFor ?? OptimizeMode.Speed; }
-            set { __pbn__OptimizeFor = value; }
+            get => __pbn__OptimizeFor ?? OptimizeMode.Speed;
+            set => __pbn__OptimizeFor = value;
         }
         public bool ShouldSerializeOptimizeFor() => __pbn__OptimizeFor != null;
         public void ResetOptimizeFor() => __pbn__OptimizeFor = null;
@@ -655,8 +668,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string GoPackage
         {
-            get { return __pbn__GoPackage ?? ""; }
-            set { __pbn__GoPackage = value; }
+            get => __pbn__GoPackage ?? "";
+            set => __pbn__GoPackage = value;
         }
         public bool ShouldSerializeGoPackage() => __pbn__GoPackage != null;
         public void ResetGoPackage() => __pbn__GoPackage = null;
@@ -666,8 +679,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool CcGenericServices
         {
-            get { return __pbn__CcGenericServices ?? false; }
-            set { __pbn__CcGenericServices = value; }
+            get => __pbn__CcGenericServices ?? false;
+            set => __pbn__CcGenericServices = value;
         }
         public bool ShouldSerializeCcGenericServices() => __pbn__CcGenericServices != null;
         public void ResetCcGenericServices() => __pbn__CcGenericServices = null;
@@ -677,8 +690,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool JavaGenericServices
         {
-            get { return __pbn__JavaGenericServices ?? false; }
-            set { __pbn__JavaGenericServices = value; }
+            get => __pbn__JavaGenericServices ?? false;
+            set => __pbn__JavaGenericServices = value;
         }
         public bool ShouldSerializeJavaGenericServices() => __pbn__JavaGenericServices != null;
         public void ResetJavaGenericServices() => __pbn__JavaGenericServices = null;
@@ -688,8 +701,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool PyGenericServices
         {
-            get { return __pbn__PyGenericServices ?? false; }
-            set { __pbn__PyGenericServices = value; }
+            get => __pbn__PyGenericServices ?? false;
+            set => __pbn__PyGenericServices = value;
         }
         public bool ShouldSerializePyGenericServices() => __pbn__PyGenericServices != null;
         public void ResetPyGenericServices() => __pbn__PyGenericServices = null;
@@ -699,8 +712,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool PhpGenericServices
         {
-            get { return __pbn__PhpGenericServices ?? false; }
-            set { __pbn__PhpGenericServices = value; }
+            get => __pbn__PhpGenericServices ?? false;
+            set => __pbn__PhpGenericServices = value;
         }
         public bool ShouldSerializePhpGenericServices() => __pbn__PhpGenericServices != null;
         public void ResetPhpGenericServices() => __pbn__PhpGenericServices = null;
@@ -710,19 +723,19 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Deprecated
         {
-            get { return __pbn__Deprecated ?? false; }
-            set { __pbn__Deprecated = value; }
+            get => __pbn__Deprecated ?? false;
+            set => __pbn__Deprecated = value;
         }
         public bool ShouldSerializeDeprecated() => __pbn__Deprecated != null;
         public void ResetDeprecated() => __pbn__Deprecated = null;
         private bool? __pbn__Deprecated;
 
         [global::ProtoBuf.ProtoMember(31, Name = @"cc_enable_arenas")]
-        [global::System.ComponentModel.DefaultValue(false)]
+        [global::System.ComponentModel.DefaultValue(true)]
         public bool CcEnableArenas
         {
-            get { return __pbn__CcEnableArenas ?? false; }
-            set { __pbn__CcEnableArenas = value; }
+            get => __pbn__CcEnableArenas ?? true;
+            set => __pbn__CcEnableArenas = value;
         }
         public bool ShouldSerializeCcEnableArenas() => __pbn__CcEnableArenas != null;
         public void ResetCcEnableArenas() => __pbn__CcEnableArenas = null;
@@ -732,8 +745,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string ObjcClassPrefix
         {
-            get { return __pbn__ObjcClassPrefix ?? ""; }
-            set { __pbn__ObjcClassPrefix = value; }
+            get => __pbn__ObjcClassPrefix ?? "";
+            set => __pbn__ObjcClassPrefix = value;
         }
         public bool ShouldSerializeObjcClassPrefix() => __pbn__ObjcClassPrefix != null;
         public void ResetObjcClassPrefix() => __pbn__ObjcClassPrefix = null;
@@ -743,8 +756,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string CsharpNamespace
         {
-            get { return __pbn__CsharpNamespace ?? ""; }
-            set { __pbn__CsharpNamespace = value; }
+            get => __pbn__CsharpNamespace ?? "";
+            set => __pbn__CsharpNamespace = value;
         }
         public bool ShouldSerializeCsharpNamespace() => __pbn__CsharpNamespace != null;
         public void ResetCsharpNamespace() => __pbn__CsharpNamespace = null;
@@ -754,8 +767,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string SwiftPrefix
         {
-            get { return __pbn__SwiftPrefix ?? ""; }
-            set { __pbn__SwiftPrefix = value; }
+            get => __pbn__SwiftPrefix ?? "";
+            set => __pbn__SwiftPrefix = value;
         }
         public bool ShouldSerializeSwiftPrefix() => __pbn__SwiftPrefix != null;
         public void ResetSwiftPrefix() => __pbn__SwiftPrefix = null;
@@ -765,8 +778,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string PhpClassPrefix
         {
-            get { return __pbn__PhpClassPrefix ?? ""; }
-            set { __pbn__PhpClassPrefix = value; }
+            get => __pbn__PhpClassPrefix ?? "";
+            set => __pbn__PhpClassPrefix = value;
         }
         public bool ShouldSerializePhpClassPrefix() => __pbn__PhpClassPrefix != null;
         public void ResetPhpClassPrefix() => __pbn__PhpClassPrefix = null;
@@ -776,8 +789,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string PhpNamespace
         {
-            get { return __pbn__PhpNamespace ?? ""; }
-            set { __pbn__PhpNamespace = value; }
+            get => __pbn__PhpNamespace ?? "";
+            set => __pbn__PhpNamespace = value;
         }
         public bool ShouldSerializePhpNamespace() => __pbn__PhpNamespace != null;
         public void ResetPhpNamespace() => __pbn__PhpNamespace = null;
@@ -787,8 +800,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string PhpMetadataNamespace
         {
-            get { return __pbn__PhpMetadataNamespace ?? ""; }
-            set { __pbn__PhpMetadataNamespace = value; }
+            get => __pbn__PhpMetadataNamespace ?? "";
+            set => __pbn__PhpMetadataNamespace = value;
         }
         public bool ShouldSerializePhpMetadataNamespace() => __pbn__PhpMetadataNamespace != null;
         public void ResetPhpMetadataNamespace() => __pbn__PhpMetadataNamespace = null;
@@ -798,8 +811,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string RubyPackage
         {
-            get { return __pbn__RubyPackage ?? ""; }
-            set { __pbn__RubyPackage = value; }
+            get => __pbn__RubyPackage ?? "";
+            set => __pbn__RubyPackage = value;
         }
         public bool ShouldSerializeRubyPackage() => __pbn__RubyPackage != null;
         public void ResetRubyPackage() => __pbn__RubyPackage = null;
@@ -832,8 +845,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool MessageSetWireFormat
         {
-            get { return __pbn__MessageSetWireFormat ?? false; }
-            set { __pbn__MessageSetWireFormat = value; }
+            get => __pbn__MessageSetWireFormat ?? false;
+            set => __pbn__MessageSetWireFormat = value;
         }
         public bool ShouldSerializeMessageSetWireFormat() => __pbn__MessageSetWireFormat != null;
         public void ResetMessageSetWireFormat() => __pbn__MessageSetWireFormat = null;
@@ -843,8 +856,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool NoStandardDescriptorAccessor
         {
-            get { return __pbn__NoStandardDescriptorAccessor ?? false; }
-            set { __pbn__NoStandardDescriptorAccessor = value; }
+            get => __pbn__NoStandardDescriptorAccessor ?? false;
+            set => __pbn__NoStandardDescriptorAccessor = value;
         }
         public bool ShouldSerializeNoStandardDescriptorAccessor() => __pbn__NoStandardDescriptorAccessor != null;
         public void ResetNoStandardDescriptorAccessor() => __pbn__NoStandardDescriptorAccessor = null;
@@ -854,8 +867,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Deprecated
         {
-            get { return __pbn__Deprecated ?? false; }
-            set { __pbn__Deprecated = value; }
+            get => __pbn__Deprecated ?? false;
+            set => __pbn__Deprecated = value;
         }
         public bool ShouldSerializeDeprecated() => __pbn__Deprecated != null;
         public void ResetDeprecated() => __pbn__Deprecated = null;
@@ -864,8 +877,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(7, Name = @"map_entry")]
         public bool MapEntry
         {
-            get { return __pbn__MapEntry.GetValueOrDefault(); }
-            set { __pbn__MapEntry = value; }
+            get => __pbn__MapEntry.GetValueOrDefault();
+            set => __pbn__MapEntry = value;
         }
         public bool ShouldSerializeMapEntry() => __pbn__MapEntry != null;
         public void ResetMapEntry() => __pbn__MapEntry = null;
@@ -887,8 +900,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(CType.String)]
         public CType Ctype
         {
-            get { return __pbn__Ctype ?? CType.String; }
-            set { __pbn__Ctype = value; }
+            get => __pbn__Ctype ?? CType.String;
+            set => __pbn__Ctype = value;
         }
         public bool ShouldSerializeCtype() => __pbn__Ctype != null;
         public void ResetCtype() => __pbn__Ctype = null;
@@ -897,8 +910,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(2, Name = @"packed")]
         public bool Packed
         {
-            get { return __pbn__Packed.GetValueOrDefault(); }
-            set { __pbn__Packed = value; }
+            get => __pbn__Packed.GetValueOrDefault();
+            set => __pbn__Packed = value;
         }
         public bool ShouldSerializePacked() => __pbn__Packed != null;
         public void ResetPacked() => __pbn__Packed = null;
@@ -908,8 +921,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(JSType.JsNormal)]
         public JSType Jstype
         {
-            get { return __pbn__Jstype ?? JSType.JsNormal; }
-            set { __pbn__Jstype = value; }
+            get => __pbn__Jstype ?? JSType.JsNormal;
+            set => __pbn__Jstype = value;
         }
         public bool ShouldSerializeJstype() => __pbn__Jstype != null;
         public void ResetJstype() => __pbn__Jstype = null;
@@ -919,8 +932,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Lazy
         {
-            get { return __pbn__Lazy ?? false; }
-            set { __pbn__Lazy = value; }
+            get => __pbn__Lazy ?? false;
+            set => __pbn__Lazy = value;
         }
         public bool ShouldSerializeLazy() => __pbn__Lazy != null;
         public void ResetLazy() => __pbn__Lazy = null;
@@ -930,8 +943,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Deprecated
         {
-            get { return __pbn__Deprecated ?? false; }
-            set { __pbn__Deprecated = value; }
+            get => __pbn__Deprecated ?? false;
+            set => __pbn__Deprecated = value;
         }
         public bool ShouldSerializeDeprecated() => __pbn__Deprecated != null;
         public void ResetDeprecated() => __pbn__Deprecated = null;
@@ -941,8 +954,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Weak
         {
-            get { return __pbn__Weak ?? false; }
-            set { __pbn__Weak = value; }
+            get => __pbn__Weak ?? false;
+            set => __pbn__Weak = value;
         }
         public bool ShouldSerializeWeak() => __pbn__Weak != null;
         public void ResetWeak() => __pbn__Weak = null;
@@ -997,8 +1010,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(2, Name = @"allow_alias")]
         public bool AllowAlias
         {
-            get { return __pbn__AllowAlias.GetValueOrDefault(); }
-            set { __pbn__AllowAlias = value; }
+            get => __pbn__AllowAlias.GetValueOrDefault();
+            set => __pbn__AllowAlias = value;
         }
         public bool ShouldSerializeAllowAlias() => __pbn__AllowAlias != null;
         public void ResetAllowAlias() => __pbn__AllowAlias = null;
@@ -1008,8 +1021,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Deprecated
         {
-            get { return __pbn__Deprecated ?? false; }
-            set { __pbn__Deprecated = value; }
+            get => __pbn__Deprecated ?? false;
+            set => __pbn__Deprecated = value;
         }
         public bool ShouldSerializeDeprecated() => __pbn__Deprecated != null;
         public void ResetDeprecated() => __pbn__Deprecated = null;
@@ -1031,8 +1044,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Deprecated
         {
-            get { return __pbn__Deprecated ?? false; }
-            set { __pbn__Deprecated = value; }
+            get => __pbn__Deprecated ?? false;
+            set => __pbn__Deprecated = value;
         }
         public bool ShouldSerializeDeprecated() => __pbn__Deprecated != null;
         public void ResetDeprecated() => __pbn__Deprecated = null;
@@ -1054,8 +1067,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Deprecated
         {
-            get { return __pbn__Deprecated ?? false; }
-            set { __pbn__Deprecated = value; }
+            get => __pbn__Deprecated ?? false;
+            set => __pbn__Deprecated = value;
         }
         public bool ShouldSerializeDeprecated() => __pbn__Deprecated != null;
         public void ResetDeprecated() => __pbn__Deprecated = null;
@@ -1077,8 +1090,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(false)]
         public bool Deprecated
         {
-            get { return __pbn__Deprecated ?? false; }
-            set { __pbn__Deprecated = value; }
+            get => __pbn__Deprecated ?? false;
+            set => __pbn__Deprecated = value;
         }
         public bool ShouldSerializeDeprecated() => __pbn__Deprecated != null;
         public void ResetDeprecated() => __pbn__Deprecated = null;
@@ -1088,8 +1101,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue(IdempotencyLevel.IdempotencyUnknown)]
         public IdempotencyLevel idempotency_level
         {
-            get { return __pbn__idempotency_level ?? IdempotencyLevel.IdempotencyUnknown; }
-            set { __pbn__idempotency_level = value; }
+            get => __pbn__idempotency_level ?? IdempotencyLevel.IdempotencyUnknown;
+            set => __pbn__idempotency_level = value;
         }
         public bool ShouldSerializeidempotency_level() => __pbn__idempotency_level != null;
         public void Resetidempotency_level() => __pbn__idempotency_level = null;
@@ -1125,8 +1138,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string IdentifierValue
         {
-            get { return __pbn__IdentifierValue ?? ""; }
-            set { __pbn__IdentifierValue = value; }
+            get => __pbn__IdentifierValue ?? "";
+            set => __pbn__IdentifierValue = value;
         }
         public bool ShouldSerializeIdentifierValue() => __pbn__IdentifierValue != null;
         public void ResetIdentifierValue() => __pbn__IdentifierValue = null;
@@ -1135,8 +1148,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(4, Name = @"positive_int_value")]
         public ulong PositiveIntValue
         {
-            get { return __pbn__PositiveIntValue.GetValueOrDefault(); }
-            set { __pbn__PositiveIntValue = value; }
+            get => __pbn__PositiveIntValue.GetValueOrDefault();
+            set => __pbn__PositiveIntValue = value;
         }
         public bool ShouldSerializePositiveIntValue() => __pbn__PositiveIntValue != null;
         public void ResetPositiveIntValue() => __pbn__PositiveIntValue = null;
@@ -1145,8 +1158,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(5, Name = @"negative_int_value")]
         public long NegativeIntValue
         {
-            get { return __pbn__NegativeIntValue.GetValueOrDefault(); }
-            set { __pbn__NegativeIntValue = value; }
+            get => __pbn__NegativeIntValue.GetValueOrDefault();
+            set => __pbn__NegativeIntValue = value;
         }
         public bool ShouldSerializeNegativeIntValue() => __pbn__NegativeIntValue != null;
         public void ResetNegativeIntValue() => __pbn__NegativeIntValue = null;
@@ -1155,8 +1168,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(6, Name = @"double_value")]
         public double DoubleValue
         {
-            get { return __pbn__DoubleValue.GetValueOrDefault(); }
-            set { __pbn__DoubleValue = value; }
+            get => __pbn__DoubleValue.GetValueOrDefault();
+            set => __pbn__DoubleValue = value;
         }
         public bool ShouldSerializeDoubleValue() => __pbn__DoubleValue != null;
         public void ResetDoubleValue() => __pbn__DoubleValue = null;
@@ -1165,8 +1178,8 @@ namespace Google.Protobuf.Reflection
         [global::ProtoBuf.ProtoMember(7, Name = @"string_value")]
         public byte[] StringValue
         {
-            get { return __pbn__StringValue; }
-            set { __pbn__StringValue = value; }
+            get => __pbn__StringValue;
+            set => __pbn__StringValue = value;
         }
         public bool ShouldSerializeStringValue() => __pbn__StringValue != null;
         public void ResetStringValue() => __pbn__StringValue = null;
@@ -1176,8 +1189,8 @@ namespace Google.Protobuf.Reflection
         [global::System.ComponentModel.DefaultValue("")]
         public string AggregateValue
         {
-            get { return __pbn__AggregateValue ?? ""; }
-            set { __pbn__AggregateValue = value; }
+            get => __pbn__AggregateValue ?? "";
+            set => __pbn__AggregateValue = value;
         }
         public bool ShouldSerializeAggregateValue() => __pbn__AggregateValue != null;
         public void ResetAggregateValue() => __pbn__AggregateValue = null;
@@ -1227,8 +1240,8 @@ namespace Google.Protobuf.Reflection
             [global::System.ComponentModel.DefaultValue("")]
             public string LeadingComments
             {
-                get { return __pbn__LeadingComments ?? ""; }
-                set { __pbn__LeadingComments = value; }
+                get => __pbn__LeadingComments ?? "";
+                set => __pbn__LeadingComments = value;
             }
             public bool ShouldSerializeLeadingComments() => __pbn__LeadingComments != null;
             public void ResetLeadingComments() => __pbn__LeadingComments = null;
@@ -1238,8 +1251,8 @@ namespace Google.Protobuf.Reflection
             [global::System.ComponentModel.DefaultValue("")]
             public string TrailingComments
             {
-                get { return __pbn__TrailingComments ?? ""; }
-                set { __pbn__TrailingComments = value; }
+                get => __pbn__TrailingComments ?? "";
+                set => __pbn__TrailingComments = value;
             }
             public bool ShouldSerializeTrailingComments() => __pbn__TrailingComments != null;
             public void ResetTrailingComments() => __pbn__TrailingComments = null;
@@ -1276,8 +1289,8 @@ namespace Google.Protobuf.Reflection
             [global::System.ComponentModel.DefaultValue("")]
             public string SourceFile
             {
-                get { return __pbn__SourceFile ?? ""; }
-                set { __pbn__SourceFile = value; }
+                get => __pbn__SourceFile ?? "";
+                set => __pbn__SourceFile = value;
             }
             public bool ShouldSerializeSourceFile() => __pbn__SourceFile != null;
             public void ResetSourceFile() => __pbn__SourceFile = null;
@@ -1286,8 +1299,8 @@ namespace Google.Protobuf.Reflection
             [global::ProtoBuf.ProtoMember(3, Name = @"begin")]
             public int Begin
             {
-                get { return __pbn__Begin.GetValueOrDefault(); }
-                set { __pbn__Begin = value; }
+                get => __pbn__Begin.GetValueOrDefault();
+                set => __pbn__Begin = value;
             }
             public bool ShouldSerializeBegin() => __pbn__Begin != null;
             public void ResetBegin() => __pbn__Begin = null;
@@ -1296,8 +1309,8 @@ namespace Google.Protobuf.Reflection
             [global::ProtoBuf.ProtoMember(4, Name = @"end")]
             public int End
             {
-                get { return __pbn__End.GetValueOrDefault(); }
-                set { __pbn__End = value; }
+                get => __pbn__End.GetValueOrDefault();
+                set => __pbn__End = value;
             }
             public bool ShouldSerializeEnd() => __pbn__End != null;
             public void ResetEnd() => __pbn__End = null;
@@ -1309,4 +1322,5 @@ namespace Google.Protobuf.Reflection
 
 }
 
-#pragma warning restore CS0612, CS1591, CS3021, IDE1006, RCS1036, RCS1057, RCS1085, RCS1192
+#pragma warning restore CS0612, CS0618, CS1591, CS3021, IDE1006, RCS1036, RCS1057, RCS1085, RCS1192
+#endregion
