@@ -5,19 +5,29 @@ using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
 using ProtoBuf.Meta;
+using ProtoBuf.unittest;
 using Xunit;
 
 namespace ProtoBuf.Issues
 {
     public class Immutables
     {
+
+        [Fact]
+        public void ImmutableArrayValidIL()
+        {
+            var model = RuntimeTypeModel.Create();
+            model.Add(typeof(ImmutableArrayTestClass));
+            model.CompileAndVerify();
+        }
+
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
         public void CanDeserialiseImmutableArray(bool autoCompile)
         {
             var testClass = new ImmutableArrayTestClass(ImmutableArray.Create("a", "b", "c"));
-            var model = TypeModel.Create();
+            var model = RuntimeTypeModel.Create();
             model.AutoCompile = autoCompile;
 
             ImmutableArrayTestClass testClassClone;
@@ -25,7 +35,9 @@ namespace ProtoBuf.Issues
             {
                 model.Serialize(ms, testClass);
                 ms.Position = 0;
+#pragma warning disable CS0618
                 testClassClone = (ImmutableArrayTestClass)model.Deserialize(ms, null, testClass.GetType());
+#pragma warning restore CS0618
             }
 
             Assert.Equal((IEnumerable<string>)testClass.Array, (IEnumerable<string>)testClassClone.Array);
@@ -37,16 +49,16 @@ namespace ProtoBuf.Issues
         public void CanDeserialiseImmutableList(bool autoCompile)
         {
             var testClass = new ImmutableListTestClass(ImmutableList.Create("a", "b", "c"));
-            var model = TypeModel.Create();
+            var model = RuntimeTypeModel.Create();
             model.AutoCompile = autoCompile;
 
             ImmutableListTestClass testClassClone;
-            using (var ms = new MemoryStream())
-            {
-                model.Serialize(ms, testClass);
-                ms.Position = 0;
-                testClassClone = (ImmutableListTestClass)model.Deserialize(ms, null, testClass.GetType());
-            }
+            using var ms = new MemoryStream();
+            model.Serialize(ms, testClass);
+            ms.Position = 0;
+#pragma warning disable CS0618
+            testClassClone = (ImmutableListTestClass)model.Deserialize(ms, null, testClass.GetType());
+#pragma warning restore CS0618
 
             Assert.Equal((IEnumerable<string>)testClass.List, (IEnumerable<string>)testClassClone.List);
         }
@@ -57,7 +69,7 @@ namespace ProtoBuf.Issues
         public void CanDeserialiseImmutableHashSet(bool autoCompile)
         {
             var testClass = new ImmutableHashSetTestClass(ImmutableHashSet.Create("a", "b", "c"));
-            var model = TypeModel.Create();
+            var model = RuntimeTypeModel.Create();
             model.AutoCompile = autoCompile;
 
             ImmutableHashSetTestClass testClassClone;
@@ -65,7 +77,9 @@ namespace ProtoBuf.Issues
             {
                 model.Serialize(ms, testClass);
                 ms.Position = 0;
+#pragma warning disable CS0618
                 testClassClone = (ImmutableHashSetTestClass)model.Deserialize(ms, null, testClass.GetType());
+#pragma warning restore CS0618
             }
 
             Assert.True(testClass.Set.SetEquals(testClassClone.Set));
@@ -77,7 +91,7 @@ namespace ProtoBuf.Issues
         public void CanDeserialiseImmutableSortedSet(bool autoCompile)
         {
             var testClass = new ImmutableSortedSetTestClass(ImmutableSortedSet.Create("a", "b", "c"));
-            var model = TypeModel.Create();
+            var model = RuntimeTypeModel.Create();
             model.AutoCompile = autoCompile;
 
             ImmutableSortedSetTestClass testClassClone;
@@ -85,7 +99,9 @@ namespace ProtoBuf.Issues
             {
                 model.Serialize(ms, testClass);
                 ms.Position = 0;
+#pragma warning disable CS0618
                 testClassClone = (ImmutableSortedSetTestClass)model.Deserialize(ms, null, testClass.GetType());
+#pragma warning restore CS0618
             }
 
             Assert.Equal((IEnumerable<string>)testClass.Set, (IEnumerable<string>)testClassClone.Set);
@@ -101,7 +117,7 @@ namespace ProtoBuf.Issues
             builder.Add("b", "2");
             builder.Add("c", "2");
             var testClass = new ImmutableDictionaryTestClass(builder.ToImmutable());
-            var model = TypeModel.Create();
+            var model = RuntimeTypeModel.Create();
             model.AutoCompile = autoCompile;
 
             ImmutableDictionaryTestClass testClassClone;
@@ -109,7 +125,9 @@ namespace ProtoBuf.Issues
             {
                 model.Serialize(ms, testClass);
                 ms.Position = 0;
+#pragma warning disable CS0618
                 testClassClone = (ImmutableDictionaryTestClass)model.Deserialize(ms, null, testClass.GetType());
+#pragma warning restore CS0618
             }
 
             Assert.Equal((IEnumerable<KeyValuePair<string, string>>)testClass.Dictionary.OrderBy(x => x.Key), (IEnumerable<KeyValuePair<string, string>>)testClassClone.Dictionary.OrderBy(x => x.Key));
@@ -125,15 +143,20 @@ namespace ProtoBuf.Issues
             builder.Add("b", "2");
             builder.Add("c", "2");
             var testClass = new ImmutableSortedDictionaryTestClass(builder.ToImmutable());
-            var model = TypeModel.Create();
+            var model = RuntimeTypeModel.Create();
             model.AutoCompile = autoCompile;
+            model.Add(typeof(ImmutableSortedDictionaryTestClass));
+
+            model.CompileAndVerify();
 
             ImmutableSortedDictionaryTestClass testClassClone;
             using (var ms = new MemoryStream())
             {
                 model.Serialize(ms, testClass);
                 ms.Position = 0;
+#pragma warning disable CS0618
                 testClassClone = (ImmutableSortedDictionaryTestClass)model.Deserialize(ms, null, testClass.GetType());
+#pragma warning restore CS0618
             }
 
             Assert.Equal((IEnumerable<KeyValuePair<string, string>>)testClass.Dictionary, (IEnumerable<KeyValuePair<string, string>>)testClassClone.Dictionary);
