@@ -1,4 +1,6 @@
-﻿using ProtoBuf.Meta;
+﻿using ProtoBuf.Internal;
+using ProtoBuf.Meta;
+using ProtoBuf.WellKnownTypes;
 using System;
 using System.IO;
 using Xunit;
@@ -10,7 +12,7 @@ namespace ProtoBuf.Perf
         [Fact]
         public void DecimalIsOptimized()
         {
-            Assert.True(BclHelpers.DecimalOptimized);
+            Assert.True(PrimaryTypeProvider.DecimalOptimized);
         }
 
         [Theory]
@@ -20,40 +22,47 @@ namespace ProtoBuf.Perf
         public void CheckDecimalLayout(bool optimized)
 #pragma warning restore xUnit1026
         {
-            bool oldVal = BclHelpers.GuidOptimized;
+            bool oldVal = PrimaryTypeProvider.GuidOptimized;
             decimal value = 1.0000000000045000000003000000M;
             try
             {
 #if DEBUG
-                BclHelpers.DecimalOptimized = optimized;
+                PrimaryTypeProvider.DecimalOptimized = optimized;
 #endif
-                using (var ms = new MemoryStream())
+                using var ms = new MemoryStream();
+                var writeState = ProtoWriter.State.Create(ms, RuntimeTypeModel.Default);
+                try
                 {
-                    using (var writer = ProtoWriter.Create(out var state, ms, RuntimeTypeModel.Default))
-                    {
-                        ProtoWriter.WriteFieldHeader(1, WireType.String, writer, ref state);
-                        BclHelpers.WriteDecimal(value, writer, ref state);
-                        writer.Close(ref state);
-                    }
-                    var hex = BitConverter.ToString(
-                        ms.GetBuffer(), 0, (int)ms.Length);
-                    Assert.Equal("0A-12-08-C0-8D-C9-B8-C0-B4-B8-E2-3E-10-DE-9C-BF-82-02-18-38", hex);
-
-                    ms.Position = 0;
-                    using(var reader = ProtoReader.Create(out var state, ms, RuntimeTypeModel.Default))
-                    {
-                        Assert.Equal(1, reader.ReadFieldHeader(ref state));
-                        var result = BclHelpers.ReadDecimal(reader, ref state);
-                        Assert.Equal(value, result);
-                        Assert.Equal(0, reader.ReadFieldHeader(ref state));
-                    }
+                    writeState.WriteFieldHeader(1, WireType.String);
+                    BclHelpers.WriteDecimal(ref writeState, value);
+                    writeState.Close();
                 }
+                finally
+                {
+                    writeState.Dispose();
+                }
+                var hex = BitConverter.ToString(
+                    ms.GetBuffer(), 0, (int)ms.Length);
+                Assert.Equal("0A-12-08-C0-8D-C9-B8-C0-B4-B8-E2-3E-10-DE-9C-BF-82-02-18-38", hex);
 
+                ms.Position = 0;
+                var readState = ProtoReader.State.Create(ms, RuntimeTypeModel.Default);
+                try
+                {
+                    Assert.Equal(1, readState.ReadFieldHeader());
+                    var result = BclHelpers.ReadDecimal(ref readState);
+                    Assert.Equal(value, result);
+                    Assert.Equal(0, readState.ReadFieldHeader());
+                }
+                finally
+                {
+                    readState.Dispose();
+                }
             }
             finally
             {
 #if DEBUG
-                BclHelpers.DecimalOptimized = oldVal;
+                PrimaryTypeProvider.DecimalOptimized = oldVal;
 #endif
             }
         }
@@ -61,7 +70,7 @@ namespace ProtoBuf.Perf
         [Fact]
         public void GuidIsOptimized()
         {
-            Assert.True(BclHelpers.GuidOptimized);
+            Assert.True(PrimaryTypeProvider.GuidOptimized);
         }
 
         [Theory]
@@ -73,27 +82,24 @@ namespace ProtoBuf.Perf
         {
             Assert.True(Guid.TryParse("12345678-2345-3456-4567-56789a6789ab", out var value));
 
-            bool oldVal = BclHelpers.GuidOptimized;
+            bool oldVal = PrimaryTypeProvider.GuidOptimized;
             try
             {
 #if DEBUG
-                BclHelpers.GuidOptimized = optimized;
+                PrimaryTypeProvider.GuidOptimized = optimized;
 #endif
-                using (var ms = new MemoryStream())
+                using var ms = new MemoryStream();
+                CheckGuidLayoutImpl(ms, value, "0A-12-09-78-56-34-12-45-23-56-34-11-45-67-56-78-9A-67-89-AB");
+                for (int i = 0; i < 100; i++)
                 {
-                    CheckGuidLayoutImpl(ms, value, "0A-12-09-78-56-34-12-45-23-56-34-11-45-67-56-78-9A-67-89-AB");
-                    for(int i = 0; i < 100; i++)
-                    {
-                        value = Guid.NewGuid();
-                        CheckGuidLayoutImpl(ms, value, null);
-                    }
+                    value = Guid.NewGuid();
+                    CheckGuidLayoutImpl(ms, value, null);
                 }
-
             }
             finally
             {
 #if DEBUG
-                BclHelpers.GuidOptimized = oldVal;
+                PrimaryTypeProvider.GuidOptimized = oldVal;
 #endif
             }
 
@@ -103,11 +109,16 @@ namespace ProtoBuf.Perf
         {
             ms.Position = 0;
             ms.SetLength(0);
-            using (var writer = ProtoWriter.Create(out var state, ms, RuntimeTypeModel.Default))
+            var state = ProtoWriter.State.Create(ms, RuntimeTypeModel.Default);
+            try
             {
-                ProtoWriter.WriteFieldHeader(1, WireType.String, writer, ref state);
-                BclHelpers.WriteGuid(value, writer, ref state);
-                writer.Close(ref state);
+                state.WriteFieldHeader(1, WireType.String);
+                BclHelpers.WriteGuid(ref state, value);
+                state.Close();
+            }
+            finally
+            {
+                state.Dispose();
             }
 
             if (expected != null)
@@ -118,12 +129,17 @@ namespace ProtoBuf.Perf
             }
 
             ms.Position = 0;
-            using (var reader = ProtoReader.Create(out var state, ms, RuntimeTypeModel.Default))
+            var readState = ProtoReader.State.Create(ms, RuntimeTypeModel.Default);
+            try
             {
-                Assert.Equal(1, reader.ReadFieldHeader(ref state));
-                var result = BclHelpers.ReadGuid(reader, ref state);
+                Assert.Equal(1, readState.ReadFieldHeader());
+                var result = BclHelpers.ReadGuid(ref readState);
                 Assert.Equal(value, result);
-                Assert.Equal(0, reader.ReadFieldHeader(ref state));
+                Assert.Equal(0, readState.ReadFieldHeader());
+            }
+            finally
+            {
+                readState.Dispose();
             }
         }
     }
