@@ -11,10 +11,9 @@
  / /   Y o u   d o   n o t   n e e d   t o   a d d   s u p p r e s s i o n s   t o   t h i s   f i l e   m a n u a l l y .  
   
  [ a s s e m b l y :   S y s t e m . D i a g n o s t i c s . C o d e A n a l y s i s . S u p p r e s s M e s s a g e ( " M i c r o s o f t . N a m i n g " ,   " C A 1 7 0 4 : I d e n t i f i e r s S h o u l d B e S p e l l e d C o r r e c t l y " ,   M e s s a g e I d   =   " B u f " ,   S c o p e   =   " n a m e s p a c e " ,   T a r g e t   =   " A q l a S e r i a l i z e r . S e r v i c e M o d e l " ) ]  
- [ a s s e m b l y :   S y s t e m . D i a g n o s t i c s . C o d e A n a l y s i s . S u p p r e s s M e s s a g e ( " M i c r o s o f t . N a m i n g " ,   " C A 1 7 0 4 : I d e n t i f i e r s S h o u l d B e S p e l l e d C o r r e c t l y " ,   M e s s a g e I d   =   " a q l a s e r i a l i z e r " ) ]  
  [ a s s e m b l y :   S y s t e m . D i a g n o s t i c s . C o d e A n a l y s i s . S u p p r e s s M e s s a g e ( " M i c r o s o f t . N a m i n g " ,   " C A 1 7 0 9 : I d e n t i f i e r s S h o u l d B e C a s e d C o r r e c t l y " ,   M e s s a g e I d   =   " a q l a s e r i a l i z e r " ) ]  
+ [ a s s e m b l y :   S y s t e m . D i a g n o s t i c s . C o d e A n a l y s i s . S u p p r e s s M e s s a g e ( " M i c r o s o f t . N a m i n g " ,   " C A 1 7 0 4 : I d e n t i f i e r s S h o u l d B e S p e l l e d C o r r e c t l y " ,   M e s s a g e I d   =   " a q l a s e r i a l i z e r " ) ]  
  [ a s s e m b l y :   S y s t e m . D i a g n o s t i c s . C o d e A n a l y s i s . S u p p r e s s M e s s a g e ( " M i c r o s o f t . N a m i n g " ,   " C A 1 7 0 4 : I d e n t i f i e r s S h o u l d B e S p e l l e d C o r r e c t l y " ,   M e s s a g e I d   =   " B u f " ,   S c o p e   =   " n a m e s p a c e " ,   T a r g e t   =   " A q l a S e r i a l i z e r . M e t a " ) ]  
- [ a s s e m b l y :   S y s t e m . D i a g n o s t i c s . C o d e A n a l y s i s . S u p p r e s s M e s s a g e ( " M i c r o s o f t . N a m i n g " ,   " C A 1 7 0 4 : I d e n t i f i e r s S h o u l d B e S p e l l e d C o r r e c t l y " ,   M e s s a g e I d   =   " B u f " ,   S c o p e   =   " n a m e s p a c e " ,   T a r g e t   =   " P r o t o B u f " ) ]  
   
   
  # e n d i f 
